# perf(ext/fetch): speed up `resp.clone()`